requirements.txt
asgiref==3.7.2
certifi==2024.2.2
channels==4.0.0
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==4.2.10
django-session-timeout==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fonttools==4.50.0
gunicorn==21.2.0
idna==3.6
importlib_resources==6.4.0
kiwisolver==1.4.5
Levenshtein==0.25.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pdfkit==1.0.0
pillow==10.2.0
psycopg==3.1.18
psycopg-binary==3.1.13
publitio==1.3.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.2
python-barcode==0.15.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-Levenshtein==0.25.0
pytz==2024.1
rapidfuzz==3.6.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
traitlets==5.14.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
XlsxWriter==3.1.9
zipp==3.18.1
qrcode==7.4.2
num2words==0.5.13
reportlab==4.0.9
